requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-RESTful==0.3.10
web3==5.31.3
requests==2.31.0
python-dotenv==1.0.0
cryptography==41.0.7
pycryptodome==3.19.0
pytest==7.4.3
pytest-flask==1.3.0
gunicorn==21.2.0 